rework database and update all DBAccess to handle this changes

+++ src/DataAccess/Search/SearchDocumentWithEvent/SearchDocumentWithEventDataAccess.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface SearchDocumentWithEventDataAccess {
     List<Integer> getDatesEvents(Integer idEvent) throws DatabaseConnectionFailedException;
-    List<Float> getQuantityItemWithSpecificEvent(int idEvent) throws DatabaseConnectionFailedException, GetQuantityItemWithSpecificEventException;
+    List<Integer> getQuantityItemWithSpecificEvent(int idEvent) throws DatabaseConnectionFailedException, GetQuantityItemWithSpecificEventException;
 }
